requirements.txt
alembic==1.13.2
blinker==1.8.2
Bootstrap-Flask==2.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Email==1.4.4
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
prettytable==3.10.2
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.25
stripe==10.10.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.0
WTForms==3.1.2
